Add support for packed bit types
Move tests to `src/ball/types/tests/main.cpp`
Add multivector by (bit/)fixed types case

+++ include/ball/types/meta/fixed/bits.h
@@ -2,6 +2,6 @@
 #	define _INCLUDE_BALL_TYPES_META_FIXED_BITS_H_
 
 typedef unsigned char bits8_t;
-typedef unsigned long long bits_t;
+typedef unsigned long bits_t;
 
 #endif // !defined( _INCLUDE_BALL_TYPES_META_FIXED_BITS_H_ )
